Move ConsoleTrace into VisitzModel
