+++ visitz/Visitz/Navigator.cs
@@ -1,4 +1,5 @@
 ﻿using Visitz.Pages;
+using VisitzModel;
 
 namespace Visitz;
