+++ visitz/Visitz/Pages/VisitzPage.xaml.cs
@@ -1,5 +1,6 @@
 using Visitz.Extensions;
 using Visitz.ViewModels;
+using VisitzModel;
 
 namespace Visitz.Pages;
