+++ visitz/Visitz/Platforms/iOS/Behaviors/SoftPageKeyboardBehavior.iOS.cs
@@ -1,6 +1,7 @@
 ﻿using CoreGraphics;
 using Foundation;
 using UIKit;
+using VisitzModel;
 
 namespace Visitz.Behaviors;
