+++ visitz/Visitz/Views/Entity/EntitySafetyAssessView.ViewModel.cs
@@ -15,6 +15,7 @@ using Visitz.Resources.Localization;
 using Visitz.Services;
 using Visitz.Storage;
 using Visitz.ViewModels;
+using VisitzModel;
 using VisitzModel.Extensions;
 using VisitzModel.Utilities;
